startForeground in the foreground
This commit fixes a nasty bug on Android, whereby location updates were halting
in the background due to an exception preventing the service making it to the
foreground.

The solution is simply to start the foreground service as early as possible.
This has the mild drawback of showing the notification whilst the app is in the
foreground, but that is a small price to pay for reliability.

close #119